Revert to 8bit turntable

+++ firmware/src/turntable.h
@@ -10,7 +10,7 @@
 #include <stdbool.h>
 
 bool turntable_init();
-uint16_t turntable_read();
+uint8_t turntable_read();
 uint16_t turntable_raw();
 void turntable_update();
 bool turntable_is_alternative();
